separate out client and server side logic
